+++ docs/conf.py
@@ -20,7 +20,7 @@ project = 'spacevader'
 year = '2019-1'
 author = 'Pratik Shivaraikar'
 copyright = '{0}, {1}'.format(year, author)
-version = release = '0.0.3'
+version = release = '0.0.4'
 
 pygments_style = 'trac'
 templates_path = ['.']